Removed some unneeded casts.
git-svn-id: http://svn.psas.pdx.edu/svn/psas/trunk/java@2199 6115d4fc-9e01-0410-8836-dad5f2c52705

+++ src/rocketview/IMUObserver.java
@@ -90,6 +90,6 @@ class IMUObserver extends JPanel implements Observer
 		}
 
 		for(int i = 0; i < data[type].length; ++i)
-			data[type][i].addPoint(msg.getTimestamp() / (float) 100, (float) msg.getData16(i));
+			data[type][i].addPoint(msg.getTimestamp() / 100f, msg.getData16(i));
 	}
 }
